requirements.txt
asgiref==3.3.1
certifi==2020.12.5
cffi==1.14.5
chardet==4.0.0
cryptography==3.4.6
defusedxml==0.7.1
dj-database-url==0.5.0
Django==3.1.7
django-allauth==0.44.0
django-appconf==1.0.4
django-compressor==2.4
django-debug-toolbar==3.2
django-markdownx==3.0.1
django-sass-processor==0.8.2
django-taggit==1.3.0
djangocms-installer==2.0.0
idna==2.10
libsass==0.20.1
Markdown==3.3.4
oauthlib==3.1.0
Pillow==8.1.2
pkg-resources==0.0.0
pycparser==2.20
PyJWT==2.0.1
python3-openid==3.2.0
pytz==2021.1
rcssmin==1.0.6
requests==2.25.1
requests-oauthlib==1.3.0
rjsmin==1.1.0
six==1.15.0
sqlparse==0.4.1
tzlocal==2.1
urllib3==1.26.4
